Improvement #30283: UI related changes in home page
Details:

Added following files in the application to display the different fonts
fontawesome-webfont.eot
fontawesome-webfont.svg
fontawesome-webfont.ttf
fontawesome-webfont.woff
fontawesome-webfont.woff2
FontAwesome.otf
Made changes in css file to create class for the left alignment
Removed the extra classes applied to the section of 'rooms' section.
Added the address of Hacienda del Mar in the contact view page
Added the following things in oceano.yml file to display the temprature
Longitude
Latitude
Consumed the api and got the temprature of Puerto Peñasco Sonora, Mexico.
Converted the following from faraday to celsius
Current Temprature
Maximum Temprature
Minimum Temprature
Fixed the issue that phone number was crossing the div element in the footer
Reduced the below and above space for the following pages
Resources
Contact Us
Reserve
Maps
FAQ
Thanking You
Changed the color of all buttons to brown
Changed the color on hover of list item
Provided target blank to open the url in new window for the following pages
All the links in Resources view page
All the links in FAQ view page
Checked the contact form whether it is working or not
Fixed the issue of 'Rooms' section that was crossing its own section into another